% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TrumpLies}
\alias{TrumpLies}
\title{Trump Lies Database}
\format{A `tbl_df`, `tbl` and `data.frame` with 4017 rows and 15 variables:
\describe{
  \item{\code{number}}{integer; Sequential lie number}
  \item{\code{date}}{Date; date of the lie}
  \item{\code{place}}{character; Place the false claim was delivered or reported}
  \item{\code{claim}}{character; Text of the false claim}
  \item{\code{infact}}{character; Text of the actual fact}
  \item{\code{slug}}{character; Brief description of the false claim}
  \item{\code{repeat}}{Factor w/ 2 levels: "no","Yes"; Was this false claim repeated?}
  \item{\code{source}}{character Web URL pointing to a story}
  \item{\code{category1}}{Factor w/ 95 levels; Content category of the false claim}
  \item{\code{category2}}{Factor w/ 101 levels; Content category of the false claim}
  \item{\code{category3}}{Factor w/ 87 levels; Content category of the false claim}
  \item{\code{category4}}{Factor w/ 48 levels; Content category of the false claim}
  \item{\code{placecategory}}{character Category of the `place`}
  \item{\code{month}}{numeric; Month of the lie, from `date` for convenience}
  \item{\code{year}}{numeric; Year of the lie, from `date` for convenience}
}}
\source{
Daniel Dale and Edward Tubb, \emph{The Toronto Star}, \url{link to Github site}
}
\usage{
TrumpLies
}
\description{
This is the complete database of false claims ("lies") by Donald Trump compiled by Daniel Dale of
       \emph{The Toronto Star} from 2017-01-20 to 2018-12-05. MORE DESCRIPTION HERE if needed.
}
\details{
DETAILS: A description of the details of codings for `slug`, `category*` and other details needed here.
         Should perhaps also give the code (somewhere) used to read in the CSV file.
}
\examples{
data(TrumpLies)
names(TrumpLies)
# do something mildly interesting here

}
\author{
Michael Friendly
}
\keyword{datasets}
